extract `WeakRefVirtualProxy` into separate file
